[CDAC] Regenerated demo for cdac_cdac_wav_gen_dma_pic32mk_mca_curiosity_pro
MH3-77739

+++ apps/cdac/cdac_wav_gen_dma/firmware/src/config/pic32mk_mca_curiosity_pro/peripheral/tmr1/plib_tmr1.c
@@ -50,6 +50,7 @@
 
 #include "device.h"
 #include "plib_tmr1.h"
+#include "interrupts.h"
 
 
 void TMR1_Initialize(void)
@@ -103,7 +104,7 @@ uint16_t TMR1_PeriodGet(void)
 
 uint16_t TMR1_CounterGet(void)
 {
-    return(TMR1);
+    return((uint16_t)TMR1);
 }
 
 uint32_t TMR1_FrequencyGet(void)
